Added the ability to put a pdf of a student's schedule into "data\new student schedules".
txtschedule_reader.py will read the pdfs and pull
schedules from them (minus the email and homeroom data- which tbh
I dont even know what it's for).

editDatabase.py now has an option to select this method, however,
the user must manually enter the email for each pdf since it's
not stored on the pdf.